Move stateMachine wrappers to shared module.
