+++ shared/domain/seasondetails/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/seasondetails/SeasonDetailsStateMachine.kt
@@ -3,14 +3,11 @@ package com.thomaskioko.tvmaniac.data.seasondetails
 import com.freeletics.flowredux.dsl.ChangedState
 import com.freeletics.flowredux.dsl.FlowReduxStateMachine
 import com.freeletics.flowredux.dsl.State
-import com.thomaskioko.tvmaniac.base.model.AppCoroutineScope
 import com.thomaskioko.tvmaniac.core.networkutil.Either
 import com.thomaskioko.tvmaniac.episodes.api.EpisodeRepository
 import com.thomaskioko.tvmaniac.seasondetails.api.SeasonDetailsRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.launch
 import me.tatarka.inject.annotations.Inject
 
 @OptIn(FlowPreview::class, ExperimentalCoroutinesApi::class)
@@ -92,33 +89,4 @@ class SeasonDetailsStateMachine constructor(
         return state.override { nextState }
     }
 }
-
-/**
- * A wrapper class around [SeasonDetailsStateMachine] handling `Flow` and suspend functions on iOS.
- */
-@Inject
-class SeasonDetailsStateMachineWrapper(
-    private val scope: AppCoroutineScope,
-    private val stateMachine: SeasonDetailsStateMachine,
-) {
-
-    fun start(stateChangeListener: (SeasonDetailsState) -> Unit) {
-        scope.main.launch {
-            stateMachine.state.collect {
-                stateChangeListener(it)
-            }
-        }
-    }
-
-    fun dispatch(action: SeasonDetailsAction) {
-        scope.main.launch {
-            stateMachine.dispatch(action)
-        }
-    }
-
-    fun cancel() {
-        scope.main.cancel()
-    }
-}
-
