+++ shared/domain/settings/src/commonMain/kotlin/com/thomaskioko/tvmaniac/settings/SettingsStateMachine.kt
@@ -1,12 +1,9 @@
 package com.thomaskioko.tvmaniac.settings
 
 import com.freeletics.flowredux.dsl.FlowReduxStateMachine
-import com.thomaskioko.tvmaniac.base.model.AppCoroutineScope
 import com.thomaskioko.tvmaniac.datastore.api.DatastoreRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.launch
 import me.tatarka.inject.annotations.Inject
 
 @OptIn(FlowPreview::class, ExperimentalCoroutinesApi::class)
@@ -62,33 +59,5 @@ class SettingsStateMachine(
                 }
             }
         }
-    }
-}
-
-/**
- * A wrapper class around [SettingsStateMachine] handling `Flow` and suspend functions on iOS.
- */
-@Inject
-class SettingsStateMachineWrapper(
-    private val scope: AppCoroutineScope,
-    private val stateMachine: SettingsStateMachine,
-) {
-
-    fun start(stateChangeListener: (SettingsState) -> Unit) {
-        scope.main.launch {
-            stateMachine.state.collect {
-                stateChangeListener(it)
-            }
-        }
-    }
-
-    fun dispatch(action: SettingsActions) {
-        scope.main.launch {
-            stateMachine.dispatch(action)
-        }
-    }
-
-    fun cancel() {
-        scope.main.cancel()
     }
 }
