+++ shared/domain/trailers/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/trailers/TrailersStateMachine.kt
@@ -3,12 +3,9 @@ package com.thomaskioko.tvmaniac.data.trailers
 import com.freeletics.flowredux.dsl.ChangedState
 import com.freeletics.flowredux.dsl.FlowReduxStateMachine
 import com.freeletics.flowredux.dsl.State
-import com.thomaskioko.tvmaniac.base.model.AppCoroutineScope
 import com.thomaskioko.tvmaniac.data.trailers.implementation.TrailerRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.launch
 import me.tatarka.inject.annotations.Inject
 
 
@@ -67,33 +64,5 @@ class TrailersStateMachine(
                 )
             }
         return state.override { nextState }
-    }
-}
-
-/**
- * A wrapper class around [TrailersStateMachine] handling `Flow` and suspend functions on iOS.
- */
-@Inject
-class TrailersStateMachineWrapper(
-    private val scope: AppCoroutineScope,
-    private val stateMachine: TrailersStateMachine,
-) {
-
-    fun start(stateChangeListener: (TrailersState) -> Unit) {
-        scope.main.launch {
-            stateMachine.state.collect {
-                stateChangeListener(it)
-            }
-        }
-    }
-
-    fun dispatch(action: TrailersAction) {
-        scope.main.launch {
-            stateMachine.dispatch(action)
-        }
-    }
-
-    fun cancel() {
-        scope.main.cancel()
     }
 }
